wip: refacto params
