+++ search-service/src/main/kotlin/com/egm/stellio/search/entity/service/EntityQueryService.kt
@@ -21,7 +21,7 @@ import com.egm.stellio.shared.model.APIException
 import com.egm.stellio.shared.model.AlreadyExistsException
 import com.egm.stellio.shared.model.ExpandedEntity
 import com.egm.stellio.shared.model.ResourceNotFoundException
-import com.egm.stellio.shared.model.param.buildGeoQuery
+import com.egm.stellio.shared.model.parameter.buildGeoQuery
 import com.egm.stellio.shared.util.Sub
 import com.egm.stellio.shared.util.buildQQuery
 import com.egm.stellio.shared.util.buildScopeQQuery
