--- shared/src/main/kotlin/com/egm/stellio/shared/model/param/GeoQueryUtils.kt
@@ -1,122 +1,0 @@
-package com.egm.stellio.shared.model.param
-
-import arrow.core.Either
-import arrow.core.left
-import arrow.core.raise.either
-import arrow.core.right
-import com.egm.stellio.shared.model.APIException
-import com.egm.stellio.shared.model.BadRequestDataException
-import com.egm.stellio.shared.model.ExpandedEntity
-import com.egm.stellio.shared.model.WKTCoordinates
-import com.egm.stellio.shared.util.JsonLdUtils
-import com.egm.stellio.shared.util.JsonLdUtils.JSONLD_VALUE
-import com.egm.stellio.shared.util.JsonLdUtils.NGSILD_GEOPROPERTY_VALUE
-import com.egm.stellio.shared.util.JsonLdUtils.expandJsonLdTerm
-import com.egm.stellio.shared.util.JsonUtils
-import com.egm.stellio.shared.util.decode
-import com.egm.stellio.shared.util.geoJsonToWkt
-
-const val GEOREL_NEAR_DISTANCE_MODIFIER = "distance"
-const val GEOREL_NEAR_MAXDISTANCE_MODIFIER = "maxDistance"
-
-private val georelNearRegex = "^near;(?:minDistance|maxDistance)==\\d+$".toRegex()
-
-fun parseGeoQueryParameters(
-    requestParams: Map<String, String>,
-    contexts: List<String>
-): Either<APIException, GeoQuery?> = either {
-    val georel = requestParams[GEO_QUERY_PARAM_GEOREL]?.decode()?.also {
-        checkGeorelParam(it).bind()
-    }
-    val geometry = requestParams[GEO_QUERY_PARAM_GEOMETRY]?.let {
-        if (GeoQuery.GeometryType.isSupportedType(it))
-            GeoQuery.GeometryType.forType(it).right()
-        else
-            BadRequestDataException("$it is not a recognized value for 'geometry' parameter").left()
-    }?.bind()
-    val coordinates = requestParams[GEO_QUERY_PARAM_COORDINATES]?.decode()?.let {
-        stringifyCoordinates(it)
-    }
-    val geoproperty = requestParams[GEO_QUERY_PARAM_GEOPROPERTY]?.let {
-        expandJsonLdTerm(it, contexts)
-    } ?: JsonLdUtils.NGSILD_LOCATION_PROPERTY
-
-    // if at least one parameter is provided, the three must be provided for the geoquery to be valid
-    val notNullGeoParameters = listOfNotNull(georel, geometry, coordinates)
-    if (notNullGeoParameters.isEmpty())
-        null
-    else if (georel == null || geometry == null || coordinates == null)
-        BadRequestDataException("Missing at least one geo parameter between 'geometry', 'georel' and 'coordinates'")
-            .left().bind<GeoQuery>()
-    else
-        GeoQuery(
-            georel = georel,
-            geometry = geometry,
-            coordinates = coordinates,
-            wktCoordinates = parseGeometryToWKT(geometry, coordinates).bind(),
-            geoproperty = geoproperty
-        )
-}
-
-fun checkGeorelParam(georel: String): Either<APIException, Unit> {
-    if (georel.startsWith(GEO_QUERY_GEOREL_NEAR)) {
-        if (!georel.matches(georelNearRegex))
-            return BadRequestDataException("Invalid expression for 'near' georel: $georel").left()
-        return Unit.right()
-    } else if (GEO_QUERY_ALL_GEORELS.any { georel == it })
-        return Unit.right()
-    else return BadRequestDataException("Invalid 'georel' parameter provided: $georel").left()
-}
-
-fun stringifyCoordinates(coordinates: Any): String =
-    when (coordinates) {
-        is String -> coordinates
-        is List<*> -> coordinates.toString()
-        else -> coordinates.toString()
-    }
-
-fun parseGeometryToWKT(
-    geometryType: GeoQuery.GeometryType,
-    coordinates: String
-): Either<APIException, WKTCoordinates> =
-    geoJsonToWkt(geometryType, coordinates)
-
-private fun prepareGeorelQuery(georel: String): Triple<String, String?, String?> =
-    if (georel.startsWith(GEO_QUERY_GEOREL_NEAR)) {
-        val comparisonParams = georel.split(";")[1].split("==")
-        if (comparisonParams[0] == GEOREL_NEAR_MAXDISTANCE_MODIFIER)
-            Triple(GEOREL_NEAR_DISTANCE_MODIFIER, "<=", comparisonParams[1])
-        else Triple(GEOREL_NEAR_DISTANCE_MODIFIER, ">=", comparisonParams[1])
-    } else Triple(georel, null, null)
-
-fun buildGeoQuery(geoQuery: GeoQuery, target: ExpandedEntity? = null): String {
-    val targetWKTCoordinates =
-        """
-        (select jsonb_path_query_first(#{TARGET}#, '$."${geoQuery.geoproperty}"."$NGSILD_GEOPROPERTY_VALUE"[0]')->>'$JSONLD_VALUE')
-        """.trimIndent()
-    val georelQuery = prepareGeorelQuery(geoQuery.georel)
-
-    return (
-        if (georelQuery.first == GEOREL_NEAR_DISTANCE_MODIFIER)
-            """
-            public.ST_Distance(
-                cast('SRID=4326;${geoQuery.wktCoordinates.value}' as public.geography), 
-                cast('SRID=4326;' || $targetWKTCoordinates as public.geography),
-                false
-            ) ${georelQuery.second} ${georelQuery.third}
-            """.trimIndent()
-        else
-            """
-            public.ST_${georelQuery.first}(
-                public.ST_GeomFromText('${geoQuery.wktCoordinates.value}'), 
-                public.ST_GeomFromText($targetWKTCoordinates)
-            ) 
-            """.trimIndent()
-        )
-        .let {
-            if (target == null)
-                it.replace("#{TARGET}#", "entity_payload.payload")
-            else
-                it.replace("#{TARGET}#", "'" + JsonUtils.serializeObject(target.members) + "'")
-        }
-}
