+++ shared/src/main/kotlin/com/egm/stellio/shared/util/GeoUtils.kt
@@ -6,7 +6,7 @@ import arrow.core.right
 import com.egm.stellio.shared.model.APIException
 import com.egm.stellio.shared.model.BadRequestDataException
 import com.egm.stellio.shared.model.WKTCoordinates
-import com.egm.stellio.shared.model.param.GeoQuery
+import com.egm.stellio.shared.model.parameter.GeoQuery
 import com.egm.stellio.shared.util.JsonUtils.deserializeObject
 import com.egm.stellio.shared.util.JsonUtils.serializeObject
 import org.locationtech.jts.io.WKTReader
