+++ subscription-service/src/main/kotlin/com/egm/stellio/subscription/service/SubscriptionService.kt
@@ -12,9 +12,9 @@ import com.egm.stellio.shared.model.ExpandedEntity
 import com.egm.stellio.shared.model.ExpandedTerm
 import com.egm.stellio.shared.model.NotImplementedException
 import com.egm.stellio.shared.model.WKTCoordinates
-import com.egm.stellio.shared.model.param.GeoQuery
-import com.egm.stellio.shared.model.param.buildGeoQuery
-import com.egm.stellio.shared.model.param.parseGeoQueryParameters
+import com.egm.stellio.shared.model.parameter.GeoQuery
+import com.egm.stellio.shared.model.parameter.buildGeoQuery
+import com.egm.stellio.shared.model.parameter.parseGeoQueryParameters
 import com.egm.stellio.shared.util.JsonLdUtils
 import com.egm.stellio.shared.util.JsonLdUtils.JSONLD_TYPE_TERM
 import com.egm.stellio.shared.util.JsonLdUtils.NGSILD_LOCATION_PROPERTY
